third_party/WebKit/LayoutTests/http/tests/push_messaging/get_registration.html

<!DOCTYPE html>
<html>
<head>
<title>Test PushManager.getRegistration()</title>
<link rel="manifest" href="resources/push_manifest.json">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../serviceworker/resources/test-helpers.js"></script>
</head>
<body>
<script>
async_test(function(test) {
    var workerUrl = 'resources/empty_worker.js';
    var workerScope = 'resources/scope/' + location.pathname;
    var swRegistration = null;
    service_worker_unregister_and_register(test, workerUrl, workerScope)
        .then(function(serviceWorkerRegistration) {
            swRegistration = serviceWorkerRegistration;
            return wait_for_state(test, swRegistration.installing, 'activated');
        })
        .then(function() {
            // If running manually, grant permission when prompted.
            if (self.testRunner)
                testRunner.setPushMessagingPermission(location.origin, true);

            assert_inherits(swRegistration.pushManager, 'getRegistration',
                            'getRegistration() should be exposed on the PushManager object');
            assert_equals(typeof(swRegistration.pushManager.getRegistration), 'function',
                          'PushManager.getRegistration() is a function.');

            return swRegistration.pushManager.getRegistration();
        })
        .then(function(pushRegistration) {
            assert_equals(pushRegistration, null,
                          "pushRegistration should be null if there is no active push registration.")
            return swRegistration.pushManager.register();
        })
        .then(function(pushRegistration) {
            previousPushRegistration = pushRegistration;
            return swRegistration.pushManager.getRegistration();
        })
        .then(function(pushRegistration) {
            assert_equals(previousPushRegistration.endPoint, pushRegistration.endPoint,
                          "Both registration objects should have the same endpoint.");
            assert_equals(previousPushRegistration.registrationId, pushRegistration.registrationId,
                          "Both registration objects should have the same registration id.");
            return pushRegistration.unregister();
        })
        .then(function(unregistered) {
            assert_true(unregistered, "unregistration was successful");
            return swRegistration.pushManager.getRegistration();
        })
        .then(function(pushRegistration) {
            assert_equals(pushRegistration, null,
                          "pushRegistration should be null after unregistering.")
            return swRegistration.pushManager.register();
        })
        .then(function(pushRegistration) {
            assert_not_equals(pushRegistration, null,
                              "Registration should have succeeded.");
            return service_worker_unregister(test, workerScope);
        })
        .then(function() {
            return swRegistration.pushManager.getRegistration();
        })
        .then(function(pushRegistration) {
            assert_equals(pushRegistration, null,
                          "After unregistration from SW, there should be no registration.");
            return service_worker_unregister_and_done(test, workerScope);
        })
        .catch(unreached_rejection(test));
});
</script>
</body>
</html>
